Apply Redux on PostCreate
