+++ nyan-forum/src/views/PostCreate/index.js
@@ -1,7 +1,9 @@
 import React from 'react';
+import { connect } from 'react-redux';
 import styled from 'styled-components';
-import { db } from '../../firebase'
 import Button from '../../components/Button';
+
+import { createPost } from './actions';
 
 class PostCreate extends React.Component {
   state = {
@@ -18,32 +20,14 @@ class PostCreate extends React.Component {
 
   handleSubmit = async (evt) => {
     evt.preventDefault();
+    const { dispatch } = this.props;
     const confirm = window.confirm('정말로 등록하시겠습니까?');
 
     if (confirm) {
       const { boardId } = this.props.match.params;
       const { author, title, content } = this.state;
-      const newPostRef = db.collection('posts').doc();
-
-      // document 생성
-      await newPostRef.set({
-        id: newPostRef.id,
-        author,
-        title,
-        content,
-        comments: []
-      })
-
-      // board의 posts 갱신 (board doc update)
-      const boardSnapshot = await db.collection('boards').doc(boardId).get();
-      console.log('boardSnapshot', boardSnapshot)
-      const postsByBoard = boardSnapshot.data().posts;
-      await db.collection('boards').doc(boardId).update({
-        posts: [
-          ...postsByBoard,
-          newPostRef.id
-        ]
-      });
+      
+      dispatch(createPost({author, title, content, boardId}));
       alert('성공적으로 등록했습니다')
       this.props.history.goBack();
     }
@@ -136,4 +120,4 @@ const Textarea = styled.textarea`
   padding: 1rem;
 `;
 
-export default PostCreate;
+export default connect()(PostCreate);
